Implemented basic repository, viewModel and tested api working

+++ app/src/main/java/com/adiupd123/beerbuzz/models/remote/BeersResponseItem.kt
@@ -2,12 +2,12 @@ package com.adiupd123.beerbuzz.models.remote
 
 data class BeersResponseItem(
     val abv: Double,
-    val attenuation_level: Int,
+    val attenuation_level: Double,
     val boil_volume: BoilVolume,
     val brewers_tips: String,
     val contributed_by: String,
     val description: String,
-    val ebc: Int,
+    val ebc: Double,
     val first_brewed: String,
     val food_pairing: List<String>,
     val ibu: Double,
@@ -17,9 +17,9 @@ data class BeersResponseItem(
     val method: Method,
     val name: String,
     val ph: Double,
-    val srm: Int,
+    val srm: Double,
     val tagline: String,
-    val target_fg: Int,
+    val target_fg: Double       ,
     val target_og: Double,
     val volume: Volume
 )
